Move FutureType functional extensions to separate files
