+++ Sources/FutureType.swift
@@ -118,94 +118,6 @@ extension FutureType {
 }
 
 extension FutureType {
-    /// Begins another asynchronous operation with the deferred value once it
-    /// becomes determined.
-    ///
-    /// `flatMap` is similar to `map`, but `transform` returns a `Deferred`
-    /// instead of an immediate value. Use `flatMap` when you want this future
-    /// to feed into another asynchronous operation. You might hear this
-    /// referred to as "chaining" or "binding".
-    ///
-    /// - parameter queue: Optional dispatch queue for starting the new
-    ///   operation from. Defaults to a global queue matching the current QoS.
-    /// - parameter transform: Start a new operation using the deferred value.
-    /// - returns: The new deferred value returned by the `transform`.
-    /// - seealso: Deferred
-    public func flatMap<NewFuture: FutureType>(upon queue: dispatch_queue_t = Self.genericQueue, _ transform: Value -> NewFuture) -> Future<NewFuture.Value> {
-        let d = Deferred<NewFuture.Value>()
-        upon(queue) {
-            transform($0).upon(queue) {
-                d.fill($0)
-            }
-        }
-        return Future(d)
-    }
-
-    /// Transforms the future once it becomes determined.
-    ///
-    /// `map` executes a transform immediately when the future's value is
-    /// determined.
-    ///
-    /// - parameter queue: Optional dispatch queue for executing the transform
-    ///   from. Defaults to a global queue matching the current QoS.
-    /// - parameter transform: Create something using the deferred value.
-    /// - returns: A new future that is filled once the reciever is determined.
-    /// - seealso: Deferred
-    public func map<NewValue>(upon queue: dispatch_queue_t = Self.genericQueue, _ transform: Value -> NewValue) -> Future<NewValue> {
-        let d = Deferred<NewValue>()
-        upon(queue) {
-            d.fill(transform($0))
-        }
-        return Future(d)
-    }
-}
-
-extension FutureType {
-    /// Composes this future with another.
-    ///
-    /// - parameter other: Any other future.
-    /// - returns: A value that becomes determined after both the reciever and
-    ///   the given future become determined.
-    /// - seealso: SequenceType.allFutures
-    public func and<OtherFuture: FutureType>(other: OtherFuture) -> Future<(Value, OtherFuture.Value)> {
-        return Future(flatMap { t in other.map { u in (t, u) } })
-    }
-    
-    /// Composes this future with others.
-    ///
-    /// - parameter one: Some other future to join with.
-    /// - parameter two: Some other future to join with.
-    /// - returns: A value that becomes determined after the reciever and both
-    ///   other futures become determined.
-    /// - seealso: SequenceType.allFutures
-    public func and<Other1: FutureType, Other2: FutureType>(one: Other1, _ two: Other2) -> Future<(Value, Other1.Value, Other2.Value)> {
-        return Future(flatMap { t in
-            one.flatMap { u in
-                two.map { v in (t, u, v) }
-            }
-        })
-    }
-    
-    /// Composes this future with others.
-    ///
-    /// - parameter one: Some other future to join with.
-    /// - parameter two: Some other future to join with.
-    /// - parameter three: Some other future to join with.
-    /// - returns: A value that becomes determined after the reciever and both
-    ///   other futures become determined.
-    /// - seealso: SequenceType.allFutures
-    public func and<Other1: FutureType, Other2: FutureType, Other3: FutureType>(one: Other1, _ two: Other2, _ three: Other3) -> Future<(Value, Other1.Value, Other2.Value, Other3.Value)> {
-        return Future(flatMap { t in
-            one.flatMap { u in
-                two.flatMap { v in
-                    three.map { w in (t, u, v, w) }
-                }
-            }
-        })
-    }
-}
-
-extension FutureType {
 
     /// A textual representation of `self`, suitable for debugging.
     public var debugDescription: String {
